added stubs for remaining FUSE functions
The read function doesn't work as expected as of right now. Some number
of bytes (a couple thousand or less) are not received when another
program reads the FUSE file. However, the bytes returned by the read
function do present a valid representation of the file, so I'm not sure
what the problem is. If the same bytes returned from read() are written
to another file on the host FS then that file is a perfect copy of the
original source. This may be a bug in the FUSE library?